all done with the project

+++ Decryptor.py
@@ -27,10 +27,10 @@ class Decrypt:
     
     def sepLines(self):
         self.fileLines=self.fileContent.split("\n")
-        print(self.fileLines)
+        # print(self.fileLines)
         for i in range(len(self.fileLines)):
             self.fileLines[i]=self.fileLines[i].split(" ")
-        print(self.fileLines)
+        # print(self.fileLines)
     
     def firstDecrypt(self):
         m=len(self.fileLines)
@@ -66,21 +66,39 @@ class Decrypt:
             if i!=m-1:
                 res+="\n"
         self.decryptedText=res
-        print("second decrypt:\n"+self.decryptedText+"\n\n")
+        # print("second decrypt:\n"+self.decryptedText+"\n\n")
 
     def decryptDataFile(self):
-        i=1
-        while os.path.exists(f"file{i}.txt"):
-            i+=1
-        
-        try:
-            fileObj=open(f"file{i}.txt","w")
-            fileObj.write(self.decryptedText)
-            file_path = os.path.abspath(fileObj.name)
-            print(f"New file created with decyphered text at :{file_path}")
-            fileObj.close()
-        except Exception as x:
-            print(f"Error in creating encrypt file:\n{x}")
+        fileChoice=input("\n\nDo you want to name the file(y/n):")
+        if fileChoice=="y":
+            while True:
+                fname=input("Enter the name of the file you want to create:")
+                fname=f"{fname}.txt"
+                if os.path.exists(fname):
+                    print("File already exists.Input another name.")
+                else:
+                    break
+            try:
+                fileObj=open(fname,"w")
+                fileObj.write(self.decryptedText)
+                file_path = os.path.abspath(fileObj.name)
+                print(f"New file created with cyphered text at :{file_path}")
+                fileObj.close()
+            except Exception as e:
+                print(e)
+
+        else:
+            i=1
+            while os.path.exists(f"file{i}.txt"):
+                i+=1
+            try:
+                fileObj=open(f"file{i}.txt","w")
+                fileObj.write(self.decryptedText)
+                file_path = os.path.abspath(fileObj.name)
+                print(f"New file created with cyphered text at :{file_path}")
+                fileObj.close()
+            except Exception as x:
+                print(f"Error in creating encrypt file:\n{x}")
         
 
     def __str__(self) -> str:
